Corrects type reading and writing for weights

+++ src/graph_binary.cpp
@@ -85,9 +85,9 @@ Graph::Graph(char *filename, char *filename_w, int type, bool renumbered) {
     ifstream finput_w;
     finput_w.open(filename_w,fstream::in | fstream::binary);
     weights.resize(nb_links_out);
-    finput_w.read((char *)&weights[0], (long)nb_links_out*sizeof(int));  
+    finput_w.read((char *)&weights[0], nb_links_out*sizeof(double));  
     weights_in.resize(nb_links_out);
-    finput_w.read((char *)&weights_in[0], (long)nb_links_in*sizeof(int));  
+    finput_w.read((char *)&weights_in[0], nb_links_in*sizeof(double));  
     cerr << "Done" << endl;
   }
